Fix compiler warnings 	modified:   Common/Source/Dialogs/dlgConfiguration.cpp 	modified:   Common/Source/Draw/ScreenLatLon.cpp

+++ Common/Source/Dialogs/dlgConfiguration.cpp
@@ -4241,7 +4241,7 @@ int ival;
   }
   wp = (WndProperty*)wf->FindByName(TEXT("prpThermalBar"));
   if (wp) {
-    if (ThermalBar != (wp->GetDataField()->GetAsInteger())) {
+    if (ThermalBar != (unsigned)(wp->GetDataField()->GetAsInteger())) {
       ThermalBar = (wp->GetDataField()->GetAsInteger());
     }
   }
